Terminando casi todos los test

+++ src/main/java/ar/edu/utn/frba/dds/Liquidacion.java
@@ -3,8 +3,7 @@ package ar.edu.utn.frba.dds;
 public class Liquidacion implements Estado {
 
     @Override
-    public float precio(float precioBase) {
-
-        return (float) (precioBase * 0.5);
+    public Double precio(Double precioBase) {
+        return (precioBase * 0.5);
     }
 }
